[fix] rolesGuard is fixed, session decorator is created

+++ src/common/base-entity.service.ts
@@ -1,4 +1,4 @@
-import { Type } from '@nestjs/common';
+import { NotFoundException, Type } from '@nestjs/common';
 import {
   DeepPartial,
   FindManyOptions,
@@ -13,21 +13,33 @@ export interface IBaseEntityService<
   Entity extends object,
   EntityDto = undefined,
 > {
-  findOne(options: FindOneOptions<Entity>): Promise<Entity>;
+  findOne(
+    options: FindOneOptions<Entity>,
+    throwError: boolean,
+  ): Promise<Entity>;
 
-  find(options: FindOptions<Entity>): Promise<Entity[]>;
+  find(options: FindOptions<Entity>, throwError: boolean): Promise<Entity[]>;
 
+  save(entities: Entity[]): Promise<Entity[]>;
   save(entity: Entity): Promise<Entity>;
 
   updateOne(
     optionsOrEntity: FindOneOptions<Entity> | Entity,
     toUpdate: DeepPartial<Entity>,
+    throwError: boolean,
   ): Promise<Entity>;
 
-  removeOne(optionsOrEntity: FindOneOptions<Entity> | Entity): Promise<void>;
+  removeOne(
+    optionsOrEntity: FindOneOptions<Entity> | Entity,
+    throwError: boolean,
+  ): Promise<void>;
 
-  remove(optionsOrEntities: FindManyOptions<Entity> | Entity[]): Promise<void>;
+  remove(
+    optionsOrEntities: FindManyOptions<Entity> | Entity[],
+    throwError: boolean,
+  ): Promise<void>;
 
+  formatToDto(entities: Entity[]): EntityDto[];
   formatToDto(entity: Entity): EntityDto;
 }
 
@@ -48,20 +60,39 @@ export abstract class BaseEntityService<
       Type<EntityDto>
     > = undefined,
   ) {}
-  async find(options: FindManyOptions<Entity>): Promise<Entity[]> {
+
+  async find(
+    options: FindManyOptions<Entity>,
+    throwError: boolean = false,
+  ): Promise<Entity[]> {
     if (Object.values(options.where).includes(undefined)) {
       throw new Error('Properties in the options.where must be defined');
     }
 
-    return await this.entityRepository.find(options);
+    const entities: Entity[] = await this.entityRepository.find(options);
+
+    if (!entities && throwError) {
+      throw new NotFoundException();
+    }
+
+    return entities;
   }
 
-  async findOne(options: FindOneOptions<Entity>): Promise<Entity> {
+  async findOne(
+    options: FindOneOptions<Entity>,
+    throwError: boolean = false,
+  ): Promise<Entity> {
     if (Object.values(options.where).includes(undefined)) {
       throw new Error('Properties in the options.where must be defined');
     }
 
-    return await this.entityRepository.findOne(options);
+    const entity: Entity = await this.entityRepository.findOne(options);
+
+    if (!entity && throwError) {
+      throw new NotFoundException();
+    }
+
+    return entity;
   }
 
   async save(entity: Entity): Promise<Entity>;
@@ -76,24 +107,34 @@ export abstract class BaseEntityService<
 
   async remove(
     optionsOrEntities: FindManyOptions<Entity> | Entity[],
+    throwError: boolean = false,
   ): Promise<void> {
-    const entity: Entity[] =
+    const entities: Entity[] =
       'where' in <object>optionsOrEntities
         ? await this.entityRepository.find(
             optionsOrEntities as FindManyOptions<Entity>,
           )
         : (optionsOrEntities as Entity[]);
 
-    await this.entityRepository.remove(entity);
+    if (!entities && throwError) {
+      throw new NotFoundException();
+    }
+
+    await this.entityRepository.remove(entities);
   }
 
   async removeOne(
     optionsOrEntity: FindOneOptions<Entity> | Entity,
+    throwError: boolean = false,
   ): Promise<void> {
     const entity: Entity =
       'where' in <object>optionsOrEntity
         ? await this.entityRepository.findOne(optionsOrEntity)
         : (optionsOrEntity as Entity);
+
+    if (!entity && throwError) {
+      throw new NotFoundException();
+    }
 
     await this.entityRepository.remove(entity);
   }
@@ -101,18 +142,29 @@ export abstract class BaseEntityService<
   async updateOne(
     optionsOrEntity: FindOneOptions<Entity> | Entity,
     toUpdate: DeepPartial<Entity>,
+    throwError: boolean = false,
   ): Promise<Entity> {
     const entity: Entity =
       'where' in <object>optionsOrEntity
         ? await this.entityRepository.findOne(optionsOrEntity)
         : (optionsOrEntity as Entity);
 
+    if (!entity && throwError) {
+      throw new NotFoundException();
+    }
+
     this.entityRepository.merge(entity, toUpdate);
 
     return this.entityRepository.save(entity);
   }
 
-  formatToDto(entity: Entity): EntityDto {
-    return new this.entityDto(entity);
+  formatToDto(entities: Entity[]): EntityDto[];
+  formatToDto(entity: Entity): EntityDto;
+  formatToDto(entity: Entity | Entity[]): EntityDto | EntityDto[] {
+    if (Array.isArray(entity)) {
+      return entity.map((entity) => new this.entityDto(entity));
+    } else {
+      return new this.entityDto(entity);
+    }
   }
 }
